Added separate controller for webpages
