+++ src/main/java/com/marcobehler/myfancypdfinvoices/web/InvoicesController.java
@@ -13,20 +13,13 @@ import org.springframework.web.bind.annotation.*;
 import java.time.LocalDateTime;
 import java.util.List;
 
-@Controller
+@RestController
 public class InvoicesController {
 
     private final InvoiceService invoiceService;
 
     public InvoicesController(InvoiceService invoiceService) {
         this.invoiceService = invoiceService;
-    }
-
-    @GetMapping("/")
-    public String getIndexPage(Model model, @RequestParam(required = false, defaultValue = "Stranger") String username) {
-        model.addAttribute("username", username);
-        model.addAttribute("currentDate", LocalDateTime.now());
-        return "index.html";
     }
 
     @GetMapping("/invoices")
@@ -37,22 +30,5 @@ public class InvoicesController {
     @PostMapping("/invoices")
     public Invoice createInvoice(@RequestBody @Valid InvoiceDto invoiceDto) {
         return invoiceService.create(invoiceDto.getUserId(), invoiceDto.getAmount());
-    }
-
-    @GetMapping("/login")
-    public String login(Model model){
-        model.addAttribute("loginForm", new LoginForm());
-        return "login.html";
-    }
-    @PostMapping("/login")
-    public String login(@ModelAttribute @Valid LoginForm loginForm, BindingResult bindingResult, Model model){
-        if (bindingResult.hasErrors()) {
-            return "login.html";
-        }
-        if (loginForm.getUsername().equals(loginForm.getPassword())) {
-            return "redirect:/";
-        }
-        model.addAttribute("invalidCredentials", "true");
-        return "login.html";
     }
 }
